아이템드롭

+++ Assets/Scripts/SO/Character_SO.cs
@@ -8,7 +8,7 @@ public class Character_SO : ScriptableObject
     public int hp;
     public int star;
     public int lv;
-    public float exp;
+    public int exp;
 
     public enum When { turnStart, battleStart, dead, selled, bought, frontAttack, damaged, lose }
     public enum What { getHp, getAtk, getBoth, getBuff, extraAtk, summon, getGold, getExp, reduceAtk }
